+++ src/main/java/org/jolokia/docker/maven/util/WaitUtil.java
@@ -3,7 +3,6 @@ package org.jolokia.docker.maven.util;
 import java.io.IOException;
 import java.util.concurrent.TimeoutException;
 
-import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpHead;
@@ -80,47 +79,39 @@ public class WaitUtil {
     public static class HttpPingChecker implements WaitChecker {
 
         private String url;
-        private Logger log;
 
         /**
          * Ping the given URL
          *
          * @param url URL to check
-         * @param log Mojo logger
-         * @param timeout
          */
-        public HttpPingChecker(String url, Logger log) {
+        public HttpPingChecker(String url) {
             this.url = url;
-            this.log = log;
         }
 
         @Override
         public boolean check() {
             try {
                 return ping();
-            } catch (ClientProtocolException exception) {
-                log.debug(String.format("wait url '%s' exception: %s", url, exception.getMessage()));
-                return false;
             } catch (IOException exception) {
-                log.debug(String.format("wait url '%s' exception: %s", url, exception.getMessage()));
                 return false;
             }
         }
 
-        private boolean ping() throws IOException, ClientProtocolException {
-            RequestConfig requestConfig = RequestConfig.custom() //
-                    .setSocketTimeout(HTTP_PING_TIMEOUT) //
-                    .setConnectTimeout(HTTP_PING_TIMEOUT) //
-                    .setConnectionRequestTimeout(HTTP_PING_TIMEOUT) //
-                    .build();
-            CloseableHttpClient httpClient = HttpClientBuilder.create() //
-                    .setDefaultRequestConfig(requestConfig) //
+        private boolean ping() throws IOException {
+            RequestConfig requestConfig =
+                    RequestConfig.custom()
+                                 .setSocketTimeout(HTTP_PING_TIMEOUT)
+                                 .setConnectTimeout(HTTP_PING_TIMEOUT)
+                                 .setConnectionRequestTimeout(HTTP_PING_TIMEOUT)
+                                 .build();
+            CloseableHttpClient httpClient = HttpClientBuilder.create()
+                    .setDefaultRequestConfig(requestConfig)
                     .build();
             try {
                 CloseableHttpResponse response = httpClient.execute(new HttpHead(url));
                 try {
                     int responseCode = response.getStatusLine().getStatusCode();
-                    log.debug(String.format("wait url '%s' response code: %s", url, responseCode));
                     return (responseCode >= 200 && responseCode <= 399);
                 } finally {
                     response.close();
@@ -136,7 +127,7 @@ public class WaitUtil {
 
     // ====================================================================================================
 
-    public static interface WaitChecker {
+    public interface WaitChecker {
         boolean check();
         void cleanUp();
     }