+++ src/main/java/org/jolokia/docker/maven/WatchMojo.java
@@ -15,24 +15,29 @@ package org.jolokia.docker.maven;/*
  * limitations under the License.
  */
 
+import static org.jolokia.docker.maven.config.WatchMode.both;
+
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.*;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
-import org.apache.maven.plugin.assembly.InvalidAssemblerConfigurationException;
-import org.apache.maven.plugin.assembly.archive.ArchiveCreationException;
-import org.apache.maven.plugin.assembly.format.AssemblyFormattingException;
 import org.codehaus.plexus.util.StringUtils;
-import org.jolokia.docker.maven.access.*;
+import org.jolokia.docker.maven.access.DockerAccess;
+import org.jolokia.docker.maven.access.DockerAccessException;
+import org.jolokia.docker.maven.access.PortMapping;
 import org.jolokia.docker.maven.assembly.AssemblyFiles;
-import org.jolokia.docker.maven.config.*;
-import org.jolokia.docker.maven.service.*;
-import org.jolokia.docker.maven.util.*;
-
-import static org.jolokia.docker.maven.config.WatchMode.both;
+import org.jolokia.docker.maven.config.ImageConfiguration;
+import org.jolokia.docker.maven.config.WatchImageConfiguration;
+import org.jolokia.docker.maven.config.WatchMode;
+import org.jolokia.docker.maven.service.QueryService;
+import org.jolokia.docker.maven.service.RunService;
+import org.jolokia.docker.maven.util.MojoParameters;
+import org.jolokia.docker.maven.util.StartOrderResolver;
 
 /**
  * Mojo for watching source code changes.
@@ -137,7 +142,7 @@ public class WatchMojo extends AbstractBuildSupporMojo {
             throws MojoExecutionException {
         final ImageConfiguration imageConfig = watcher.getImageConfiguration();
         final String name = imageConfig.getName();
-        final AssemblyFiles files = getAssemblyFiles(name, imageConfig.getBuildConfiguration(), mojoParameters);
+        final AssemblyFiles files = serviceHub.getBuildService().getAssemblyFiles(name, imageConfig, mojoParameters);
         return new Runnable() {
             @Override
             public void run() {
@@ -184,16 +189,6 @@ public class WatchMojo extends AbstractBuildSupporMojo {
             }
         };
     }
-
-    private AssemblyFiles getAssemblyFiles(String name, BuildImageConfiguration buildConfiguration, MojoParameters mojoParameters)
-            throws MojoExecutionException {
-        try {
-            return dockerAssemblyManager.getAssemblyFiles(name, buildConfiguration, mojoParameters);
-        } catch (InvalidAssemblerConfigurationException | ArchiveCreationException | AssemblyFormattingException  e) {
-            throw new MojoExecutionException("Cannot extract assembly files for image " + name + ": " + e,e);
-        }
-    }
-
 
     private void restartContainer(ImageWatcher watcher) throws DockerAccessException {
         // Stop old one