+++ src/main/java/org/jolokia/docker/maven/config/handler/property/ConfigKey.java
@@ -35,6 +35,7 @@ public enum ConfigKey {
     BIND,
     CAP_ADD,
     CAP_DROP,
+    CLEANUP,
     CMD,
     DOMAINNAME,
     DNS,