+++ doc/changelog.md
@@ -2,6 +2,7 @@
 
 * **0.13.2**
   - "run" directives can be added to the Dockerfile (#191)
+  - Support user information in wait URL (#211)
   
 * **0.13.1**
   - Allow autoPull to be forced on docker:build and docker:start (#96)